+++ doc/changes.rst
@@ -7,6 +7,7 @@ Change Log
 
 * Adds redrock, surveysim, desisurvey, and healpy to dependency version checks.
 * Adds redrock and surveysim to known products for installation.
+* Fix team name in license file.
 
 1.9.7 (2017-09-29)
 ------------------